+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
 - **New:** Seed for conversation message threadmark (#39)
 - **New:** Seed for discouraged IP addresses (#41)
 - **New:** Seed for banned users (#43)
+- **New:** Seed for IP address bans (#44)
 - **Fix:** Threads cannot be seeded (#36)
 
 ## 1.1.0 Alpha 3 (`1010013`)