+++ CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
 - **New:** Seed for IP address bans (#44)
 - **New:** Seed for rejected users (#42)
 - **New:** Seed for server error logs (#45)
+- **New:** Seed for spam trigger log (#46)
 - **Fix:** Threads cannot be seeded (#36)
 
 ## 1.1.0 Alpha 3 (`1010013`)